#23, #18 , #8 , #26
#23, #18 , #8
- #23 : Add new reflection to attributes.js
- #18 : Review code
- #8 : Try to make the code more readable.
- #26 : Fix JavaScript braces with good practice

+++ src/scripts/styleManipulation/trimSecond.js
@@ -1,4 +1,4 @@
 ﻿$.animateJs.styleManipulation.trimSecond = function (text) {
-    var number = parseInt(text, 10);
+    var number = parseFloat(text, 10);
     return number;
 }
